Add option to search by id
    - Add id param to controller & service
    - Update documentation (swagger, readme)
    - Add test

    - Move formatting functions into helper function to keep a cleaner
    codebase
