+++ README.md
@@ -1,6 +1,6 @@
 # Location Picker Smart Widget BFF (Node)
 
-This is a Node.js backend service library to create a BFF service for the Location Picker Smart Widget. The Location Picker widget provides a picker field to choose a street, street address or point of interest from GIS sources. This service is matched by a [corresponding UI](https://github.com/digipolisantwerp/location-picker_widget_angular).
+This is a Node.js backend service library to create a BFF service for the Location Picker Smart Widget. The Location Picker widget provides a picker field to choose a street, street address, addressID or point of interest from GIS sources. This service is matched by a [corresponding UI](https://github.com/digipolisantwerp/location-picker_widget_angular).
 
 There is a **demo service**, see below for instructions on running it.
 
@@ -87,8 +87,9 @@ The UI demo app expects the service to run on port 9999.
 
 The service implements the following first protocol:
 
-- GET /path/to/endpoint?search=...&types=...
+- GET /path/to/endpoint?search=...&types=...&id=..
 - search = the text that the user typed on which to match
+- id = the id of the location. This overrides search. (Optional)
 - types = types to query for
   - Possible types are `street` (street names), `number` (address excluding bus) and `poi` (point of interest)- Comma-separated, default value is `street,number,poi`
 - result = JSON-encoded array of [LocationItem](src/types.ts) objects
